+++ CHANGELOG.md
@@ -58,4 +58,4 @@ package of the `eosc` library.
 
 * The `--auth-key` and `--auth-file` flags of the `system newaccount` command were removed, and changed to positional arguemnts, using the short-form autority syntax. The auth file support is removed, as the short-form syntax absorbs all of its functionality, in a much more terse way.
 
-* The `--setpriv` flag on `system newaccount` was replaced by the `--add-actions` option `setpriv`.
+* The `--setpriv` flag on `system newaccount` was replaced by the `--additional-actions` option `setpriv`.